+++ Docs/Whats_New.md
@@ -30,11 +30,16 @@ Released TBD
 - Fix issue where mapped host-coherent device memory not updated from image contents on *macOS*.
 - Fix small memory leak when setting swapchain color space.
 - Fix new and unexpected App Store failure on newly deprecated color space values.
+- Include vertex attribute size when testing whether attribute offset exceeds stride.
+- Add `MVKPhysicalDeviceMetalFeatures::vertexStrideAlignment` to track Metal vertex binding stride alignment.
 - Remove use of `@available()` directive as it was causing issues in some build environments.
 - Refactor **MoltenVK** *Xcode* build architectures.
 - Demo `API-Samples generateSPIRVShaders` no longer builds `MoltenVKShaderController` tool.
+- Update `VK_MVK_MOLTENVK_SPEC_VERSION` to `27`.
 - Update to latest SPIRV-Cross version:
 	- MSL: Workaround broken scalar access chain behavior in LLVM IR / AIR
+
+
 
 MoltenVK 1.0.43
 ---------------